README.md: Expand the expected usage description
This also removes the expectation of the sanity tests, mainly because
they never were created and were canned before being created because
they were originally invisioned to be CI tests, which would fail PRs the
moment a comment was made to grpc-proto.

We are instead using scripts and other solutions to keep the chances of
manual changes low, while only using a single copy of the protos
in Google's internal build. Since repositories sync to Google-internal
frequently, manual changes that "get through the cracks" and break
functionality should still be discovered promptly.

Not all of that is in the README, though, because it proved hard to
integrate such that the prose was still short and to the point, while
also not being confusing to non-Google-devs. And I don't actually expect
many Google-devs would read it anyway...