Specify new semantics for leaving NameMatcher.key empty
This field can be omitted or set to an empty string for the
HttpKeyBuilder uses, in which case the matcher will match the
specified header name or query parameter, but not update the key_map.
This brings it in line with other aspects of the matcher model by
decoupling matching from key generation.